python-service/app/requirements.txt
beautifulsoup4 == 4.11.1
json5 == 0.9.10
jsonfy == 0.4
jsonpointer == 2.3
jsonschema == 4.21.1
jsonschema-specifications == 2023.12.1
mysql-connector-python == 8.3.0
selenium == 4.17.2
twilio == 8.11.1
gunicorn == 21.2.0
webdriver-manager== 4.0.2
python-dotenv == 1.0.1
boto3 == 1.34.103
websockets==12.0 
flask == 3.0.0
scikit-learn == 1.5.1
gspread == 6.1.4
spacy == 3.8.4